Split Migrate and Snapshot Procedure Tests
